stop process for no hits

+++ branches/dev-sigurd-2/phpgwapi/inc/class.acl.inc.php
@@ -1054,6 +1054,10 @@
 				$locations[] = $this->_db->f('location_id');
 			}
 
+			if(!$locations)
+			{
+				return;
+			}
 			$location_filter = implode(',', $locations);
 
 			$sql = 'DELETE FROM phpgw_acl'
